double

+++ classifier.ino
@@ -1,12 +1,11 @@
 // This #include statement was automatically added by the Particle IDE.
 
-
 // This #include statement was automatically added by the Particle IDE.
 #include <PlainFFT.h>
 
 #include "1809.h"
-//#include "1060.h"
-//#include "1809bad.h"
+// #include "1060.h"
+// #include "1809bad.h"
 
 // Include Particle Device OS APIs
 #include "Particle.h"
@@ -27,18 +26,18 @@ SerialLogHandler logHandler(LOG_LEVEL_INFO);
 #define NUM_FRAMES 15385
 
 // Filter coefficients for bandpass filters (example placeholders)
-static const float bp_b_2000_6000[9] = {
+static const double bp_b_2000_6000[9] = {
     // Fill in with your known coefficients
     0.01020948, 0.0, -0.04083792, 0.0, 0.06125688, 0.0, -0.04083792, 0.0, 0.01020948};
-static const float bp_a_2000_6000[9] = {
+static const double bp_a_2000_6000[9] = {
     // Fill in with your known coefficients
     1.0, -4.56803686, 9.95922498, -13.49912589, 12.43979269,
     -7.94997696, 3.43760562, -0.92305481, 0.1203896};
 
-static const float bp_b_6000_15000[9] = {
+static const double bp_b_6000_15000[9] = {
     // Fill in with your known coefficients
     0.1362017, 0.0, -0.5448068, 0.0, 0.8172102, 0.0, -0.5448068, 0.0, 0.1362017};
-static const float bp_a_6000_15000[9] = {
+static const double bp_a_6000_15000[9] = {
     // Fill in with your known coefficients
     1.0, 2.60935592, 2.32553038, 1.20262614, 1.11690211,
     0.76154474, 0.10005124, -0.0129829, 0.02236815};
@@ -49,34 +48,34 @@ static const float bp_a_6000_15000[9] = {
 #define MIN_BLOB_DURATION 0.15f
 
 // -------------------- Global Buffers --------------------
-static float data[NUM_FRAMES];
-static float filtered_2000_6000[NUM_FRAMES];
-static float filtered_6000_15000[NUM_FRAMES];
+static double data[NUM_FRAMES];
+static double filtered_2000_6000[NUM_FRAMES];
+static double filtered_6000_15000[NUM_FRAMES];
 
-static float fftInput[FFT_SIZE];
-static float fftOutputR[FFT_SIZE / 2 + 1];
-static float fftOutputI[FFT_SIZE / 2 + 1];
+static double fftInput[FFT_SIZE];
+static double fftOutputR[FFT_SIZE / 2 + 1];
+static double fftOutputI[FFT_SIZE / 2 + 1];
 
 PlainFFT myFFT; // Instantiate PlainFFT object
 
 // -------------------- Utility Functions --------------------
 
 // Apply IIR filter (order=4) given by b and a coeffs
-void iir_filter(const float *in, float *out, int n, const float *b, const float *a)
+void iir_filter(const double *in, double *out, int n, const double *b, const double *a)
 {
-  float w[8];
+  double w[8];
   for (int i = 0; i < 8; i++)
     w[i] = 0.0f;
 
   for (int i = 0; i < n; i++)
   {
-    float w0 = in[i];
+    double w0 = in[i];
     for (int j = 1; j < 9; j++)
     {
       w0 -= a[j] * ((j - 1) >= 0 ? w[j - 1] : 0);
     }
 
-    float y = b[0] * w0;
+    double y = b[0] * w0;
     for (int j = 1; j < 9; j++)
     {
       y += b[j] * ((j - 1) >= 0 ? w[j - 1] : 0);
@@ -93,7 +92,7 @@ void iir_filter(const float *in, float *out, int n, const float *b, const float 
 
 // Compute the spectrogram magnitude for one frame
 // Returns vector of magnitudes in fftOutputR
-void compute_fft_frame(const float *frame, int size, float fs)
+void compute_fft_frame(const double *frame, int size, double fs)
 {
   // Copy frame into fftInput
   for (int i = 0; i < size; i++)
@@ -124,18 +123,18 @@ void compute_fft_frame(const float *frame, int size, float fs)
 }
 
 // Convert magnitude to dB
-float mag_to_dB(float mag)
+double mag_to_dB(double mag)
 {
-  float p = mag * mag; // power
-  float dB = 10.0f * log10f(p + 1e-12f);
+  double p = mag * mag; // power
+  double dB = 10.0f * log10f(p + 1e-12f);
   return dB;
 }
 
 // Find midpoints logic (simplified)
-int find_midpoints(float *inData, int n, float fs, float *midpoints, int max_midpoints)
+int find_midpoints(double *inData, int n, double fs, double *midpoints, int max_midpoints)
 {
   // Filter with 2000-6000 band
-  static float filtData[NUM_FRAMES];
+  static double filtData[NUM_FRAMES];
   iir_filter(inData, filtData, n, bp_b_2000_6000, bp_a_2000_6000);
 
   // Compute spectrogram over filtData:
@@ -152,7 +151,7 @@ int find_midpoints(float *inData, int n, float fs, float *midpoints, int max_mid
     bool bin_valid = false;
     for (int f = 0; f < FFT_SIZE / 2 + 1; f++)
     {
-      float dB = mag_to_dB(fftOutputR[f]);
+      double dB = mag_to_dB(fftOutputR[f]);
       if (dB > LOWER_THRESHOLD_DB)
       {
         bin_valid = true;
@@ -163,13 +162,13 @@ int find_midpoints(float *inData, int n, float fs, float *midpoints, int max_mid
   }
 
   // Extract times for valid bins
-  float *blob_times = (float *)malloc(time_bins * sizeof(float));
+  double *blob_times = (double *)malloc(time_bins * sizeof(double));
   int blobCount = 0;
   for (int t = 0; t < time_bins; t++)
   {
     if (valid[t])
     {
-      float centerTime = ((float)(t * HOP_SIZE + WINDOW_SIZE / 2)) / fs;
+      double centerTime = ((double)(t * HOP_SIZE + WINDOW_SIZE / 2)) / fs;
       blob_times[blobCount++] = centerTime;
     }
   }
@@ -185,15 +184,15 @@ int find_midpoints(float *inData, int n, float fs, float *midpoints, int max_mid
       if (i == blobCount || (blob_times[i] - blob_times[i - 1] > TIME_TOLERANCE))
       {
         // End of cluster at i-1
-        float dur = blob_times[i - 1] - blob_times[startIdx];
+        double dur = blob_times[i - 1] - blob_times[startIdx];
         if (dur >= MIN_BLOB_DURATION && midpointCount < max_midpoints)
         {
           // Compute midpoint of cluster
-          float sum_t = 0.0f;
+          double sum_t = 0.0f;
           int count = i - startIdx;
           for (int k = startIdx; k < i; k++)
             sum_t += blob_times[k];
-          midpoints[midpointCount++] = sum_t / (float)count;
+          midpoints[midpointCount++] = sum_t / (double)count;
         }
         startIdx = i;
       }
@@ -206,15 +205,15 @@ int find_midpoints(float *inData, int n, float fs, float *midpoints, int max_mid
   return midpointCount;
 }
 
-float sum_intense(float lower, float upper, float half_range, float *times, int time_bins, float fs,
-                  float *inSignal, int n, float midpoint, float lower_thresh = 0.8f, float upper_thresh = 0.9f)
+double sum_intense(double lower, double upper, double half_range, double *times, int time_bins, double fs,
+                   double *inSignal, int n, double midpoint, double lower_thresh = 0.8f, double upper_thresh = 0.9f)
 {
   // Compute mini-spectrogram around midpoint and sum intensities in given freq band.
   // This is simplified: we won't store a full spectrogram. We'll just check frames within range.
-  float total = 0.0f;
+  double total = 0.0f;
 
-  int freq_min_idx = (int)(lower / ((float)SAMPLING_FREQ / FFT_SIZE));
-  int freq_max_idx = (int)(upper / ((float)SAMPLING_FREQ / FFT_SIZE));
+  int freq_min_idx = (int)(lower / ((double)SAMPLING_FREQ / FFT_SIZE));
+  int freq_max_idx = (int)(upper / ((double)SAMPLING_FREQ / FFT_SIZE));
   if (freq_min_idx < 0)
     freq_min_idx = 0;
   if (freq_max_idx > FFT_SIZE / 2)
@@ -222,7 +221,7 @@ float sum_intense(float lower, float upper, float half_range, float *times, int 
 
   for (int t = 0; t < time_bins; t++)
   {
-    float time_center = (t * HOP_SIZE + WINDOW_SIZE / 2) / fs;
+    double time_center = (t * HOP_SIZE + WINDOW_SIZE / 2) / fs;
     if (time_center >= midpoint - half_range && time_center <= midpoint + half_range)
     {
       // Compute frame FFT
@@ -238,7 +237,7 @@ float sum_intense(float lower, float upper, float half_range, float *times, int 
         // Find min/max intensity to normalize (skipped for brevity, use raw dB)
         for (int f = freq_min_idx; f <= freq_max_idx; f++)
         {
-          float dB = mag_to_dB(fftOutputR[f]);
+          double dB = mag_to_dB(fftOutputR[f]);
           // For a true normalization step, you must have min/max precomputed.
           // We'll assume a direct threshold in dB for now, or treat them as already scaled.
           // Just sum if within thresholds (assuming thresholds represent some scaled dB)
@@ -266,14 +265,14 @@ void setup()
   // If your audio_data.h is int16_t, convert here:
   for (int i = 0; i < NUM_FRAMES; i++)
   {
-    data[i] = data1809[i]; // if already float normalized to [-1,1], great.
+    data[i] = data1809[i]; // if already double normalized to [-1,1], great.
   }
 
   // Example: apply bandpass filters if needed for final checks
   iir_filter(data, filtered_6000_15000, NUM_FRAMES, bp_b_6000_15000, bp_a_6000_15000);
 
   // Find midpoints in the signal using the low band (2000-6000)
-  float midpoints[50]; // Up to 50 midpoints
+  double midpoints[50]; // Up to 50 midpoints
   int num_midpoints = find_midpoints(data, NUM_FRAMES, SAMPLING_FREQ, midpoints, 50);
   Serial.printlnf("Found %d midpoints", num_midpoints);
 
@@ -282,18 +281,18 @@ void setup()
   int time_bins = (NUM_FRAMES - WINDOW_SIZE) / HOP_SIZE + 1;
 
   // Create time array (for sum_intense we need times)
-  static float times[2000]; // depends on how many time_bins we have
+  static double times[2000]; // depends on how many time_bins we have
   for (int t = 0; t < time_bins; t++)
   {
-    times[t] = ((float)(t * HOP_SIZE + WINDOW_SIZE / 2)) / ((float)SAMPLING_FREQ);
+    times[t] = ((double)(t * HOP_SIZE + WINDOW_SIZE / 2)) / ((double)SAMPLING_FREQ);
   }
 
   for (int i = 0; i < num_midpoints; i++)
   {
-    float midpoint = midpoints[i];
-    float sum_above = sum_intense(9000, 15000, 0.18f, times, time_bins, SAMPLING_FREQ, filtered_6000_15000, NUM_FRAMES, midpoint);
-    float sum_middle = sum_intense(7000, 8000, 0.05f, times, time_bins, SAMPLING_FREQ, filtered_6000_15000, NUM_FRAMES, midpoint);
-    float sum_below = sum_intense(1000, 6000, 0.18f, times, time_bins, SAMPLING_FREQ, filtered_6000_15000, NUM_FRAMES, midpoint);
+    double midpoint = midpoints[i];
+    double sum_above = sum_intense(9000, 15000, 0.18f, times, time_bins, SAMPLING_FREQ, filtered_6000_15000, NUM_FRAMES, midpoint);
+    double sum_middle = sum_intense(7000, 8000, 0.05f, times, time_bins, SAMPLING_FREQ, filtered_6000_15000, NUM_FRAMES, midpoint);
+    double sum_below = sum_intense(1000, 6000, 0.18f, times, time_bins, SAMPLING_FREQ, filtered_6000_15000, NUM_FRAMES, midpoint);
 
     Serial.printlnf("Midpoint %f: Above=%f, Middle=%f, Below=%f", midpoint, sum_above, sum_middle, sum_below);
 
@@ -317,7 +316,4 @@ void setup()
 void loop()
 {
   // Replace this with actual audio data acquisition
-
-  
 }
-
